De-emphasising most of the multiplayer code. It kinda works, but state sync is understandably complex. I'll probably rip most of it out and just have the requirement that both games start at the same time and stay in sync.
To do otherwise would require refactoring the main game to allow for
state synchronisation, incl. probably centralising game state machine
at a single point to allow for rerouting.  I don't really want to do
that; it's not that hard, but it goes against the simple structure for
the game.

Or, I could easily fork this so there's a more complex version with
multiplayer, or the simple version without.

Haven't decided yet.  Anyway, next thing to do is just to continue
commenting the code.